# runtime/chan.go: improve closed channel receive performance