--- a/src/runtime/chan_test.go
+++ b/src/runtime/chan_test.go
@@ -1132,6 +1132,8 @@ func BenchmarkReceiveDataFromClosedChan(b *testing.B) {
 		ch <- struct{}{}
 	}
 	close(ch)
+
+	b.ResetTimer()
 	for range ch {
 	}
 }